fix() : Fix change LottoResult class to LottoViewResolver includes viewer as field
