+++ src/main/java/lotto/view/Viewer.java
@@ -1,8 +1,5 @@
 package lotto.view;
 
-import lotto.data.Lottos;
-
 public interface Viewer {
-    void view(Lottos lottos);
-    void viewResult(LottoResult result);
+    void view(String message);
 }
